Move DarknessFilter to the screen instead of the header
